split network building
