+++ espnet2/asr/custom/build_encoder.py
@@ -1,4 +1,4 @@
-"""Set of methods to create customizable network."""
+"""Set of methods to build customizable encoder."""
 
 from typing import Any
 from typing import Dict
@@ -11,27 +11,15 @@ from espnet.nets.pytorch_backend.conformer.convolution import ConvolutionModule
 from espnet.nets.pytorch_backend.conformer.encoder_layer import (
     EncoderLayer as ConformerEncoderLayer,  # noqa: H301
 )
-from espnet.nets.pytorch_backend.nets_utils import get_activation
 from espnet.nets.pytorch_backend.transducer.vgg2l import VGG2L
-from espnet.nets.pytorch_backend.transformer.attention import (
-    MultiHeadedAttention,  # noqa: H301
-    RelPositionMultiHeadedAttention,  # noqa: H301
-)
-from espnet.nets.pytorch_backend.transformer.embedding import (
-    PositionalEncoding,  # noqa: H301
-    ScaledPositionalEncoding,  # noqa: H301
-    RelPositionalEncoding,  # noqa: H301
-)
 from espnet.nets.pytorch_backend.transformer.encoder_layer import (
     EncoderLayer as TransformerEncoderLayer,  # noqa: H301
 )
-from espnet.nets.pytorch_backend.transformer.multi_layer_conv import Conv1dLinear
-from espnet.nets.pytorch_backend.transformer.multi_layer_conv import MultiLayeredConv1d
-from espnet.nets.pytorch_backend.transformer.positionwise_feed_forward import (
-    PositionwiseFeedForward,  # noqa: H301
-)
 from espnet.nets.pytorch_backend.transformer.repeat import MultiSequential
 from espnet.nets.pytorch_backend.transformer.subsampling import Conv2dSubsampling
+from espnet2.asr.custom.utils import get_positional_encoding_class
+from espnet2.asr.custom.utils import get_positionwise_classes
+from espnet2.asr.custom.utils import get_self_attention_class
 
 
 def verify_layers_io(input_size: int, architecture: List[Dict[str, Any]]):
@@ -39,7 +27,7 @@ def verify_layers_io(input_size: int, architecture: List[Dict[str, Any]]):
 
     Args:
         input_size: Input layer size
-        architecture: Network architecture configuration
+        architecture: Encoder architecture configuration
 
     Returns:
         (): if architecture is valid, return output size
@@ -89,94 +77,18 @@ def verify_layers_io(input_size: int, architecture: List[Dict[str, Any]]):
     return check_io[-1][1]
 
 
-def get_positionwise_classes(
-    positionwise_layer_type: str,
-    positionwise_activation_type: str,
-) -> torch.nn.Module:
-    """Get positionwise layer related classes.
-
-    Args:
-        positionwise_layer_type: Type of positionwise layer
-        positionwise_activation_type: Type of positiontwise layer activation
-
-    Return:
-        positionwise_layer: Positionwise layer class
-        positionwise_activation: Positionwise activation class
-
-    """
-    if positionwise_layer_type == "linear":
-        positionwise_layer = PositionwiseFeedForward
-    elif positionwise_layer_type == "conv1d":
-        positionwise_layer = MultiLayeredConv1d
-    elif positionwise_layer_type == "conv1d-linear":
-        positionwise_layer = Conv1dLinear
-    else:
-        raise NotImplementedError(
-            "'positionwise_layer_type' argument only support 'linear' yet."
-        )
-
-    positionwise_activation = get_activation(positionwise_activation_type)
-
-    return positionwise_layer, positionwise_activation
-
-
-def get_positional_encoding_class(positional_encoding_type: str) -> torch.nn.Module:
-    """Get positional encoding class.
-
-    Args:
-        positional_encoding_type: Type of positional encoding
-
-    Returns:
-        positional_encoding_class: Positional encoding class
-
-    """
-    if positional_encoding_type == "abs_pos":
-        positional_encoding_class = PositionalEncoding
-    elif positional_encoding_type == "scaled_abs_pos":
-        positional_encoding_class = ScaledPositionalEncoding
-    elif positional_encoding_type == "rel_pos":
-        positional_encoding_class = RelPositionalEncoding
-    else:
-        raise NotImplementedError(
-            "Invalid positional_encoding_type" + positional_encoding_type
-        )
-
-    return positional_encoding_class
-
-
-def get_self_attention_class(self_attention_type: str) -> torch.nn.Module:
-    """Get self-attention class.
-
-    Args:
-        self_attention_type: Type of self-attention
-
-    Returns:
-        self_attention_class: Self-attention class
-
-    """
-    if self_attention_type == "self_attn":
-        self_attn_class = MultiHeadedAttention
-    elif self_attention_type == "rel_selfattn":
-        self_attn_class = RelPositionMultiHeadedAttention
-    else:
-        raise NotImplementedError(
-            "self_attn_type should be either 'self_attn' or 'rel_self_attn'"
-        )
-
-    return self_attn_class
-
-
-def create_transformer_layer(
+def build_transformer_layer(
     output_size: int,
     linear_units: int,
     attention_heads: int,
+    self_attention_layer_type: str = "selfattn",
     positionwise_layer_type: str = "linear",
     positionwise_activation_type: str = "relu",
     dropout_rate: float = 0.0,
     positionwise_dropout_rate: float = 0.0,
     attention_dropout_rate: float = 0.0,
 ) -> torch.nn.Module:
-    """Create transformer layer.
+    """Build transformer layer.
 
     Args:
         output_size:
@@ -204,15 +116,22 @@ def create_transformer_layer(
             "Transformer layer: positionwise_layer_type should be 'linear'."
         )
 
+    self_att_layer = get_self_attention_class(self_attention_layer_type)
+    self_att_layer_args = (
+        attention_heads,
+        output_size,
+        attention_dropout_rate,
+    )
+
     return lambda: TransformerEncoderLayer(
         output_size,
-        MultiHeadedAttention(attention_heads, output_size, attention_dropout_rate),
+        self_att_layer(*self_att_layer_args),
         pw_layer(*pw_layer_args),
         dropout_rate,
     )
 
 
-def create_conformer_layer(
+def build_conformer_layer(
     output_size: int,
     linear_units: int,
     attention_heads: int,
@@ -229,7 +148,7 @@ def create_conformer_layer(
     positionwise_dropout_rate: float = 0.0,
     attention_dropout_rate: float = 0.0,
 ) -> torch.nn.Module:
-    """Create transformer layer.
+    """Build transformer layer.
 
     Args:
         output_size:
@@ -280,7 +199,7 @@ def create_conformer_layer(
     )
 
 
-def create_input_layer(
+def build_input_layer(
     input_size: int,
     output_size: int = 320,
     layer_type: str = "embed",
@@ -289,7 +208,7 @@ def create_input_layer(
     positional_encoding_dropout_rate: float = 0.0,
     padding_idx: int = -1,
 ) -> torch.nn.Module:
-    """Create input layer.
+    """Build input layer.
 
     Args:
         input_size: Input dimension
@@ -339,28 +258,24 @@ def create_input_layer(
     return embed
 
 
-def create_network(
-    architecture_type: str,
+def build_encoder(
     input_size: int,
     architecture: List[Dict[str, Any]],
     repeat: int = 0,
     padding_idx: int = -1,
 ) -> Union[torch.nn.Module, MultiSequential, int]:
-    """Create network based on architecture specification.
+    """Build encoder based on architecture specification.
 
     Args:
         input_size: Input dimension
-        architecture_part: Either 'encoder' or 'decoder'
-        architecture: Network architecture configuration
+        architecture: Encoder architecture configuration
         repeat: Number of times specified architecture should be repeated
         padding_idx: Index for embedding padding
 
     """
-    assert architecture_type == "encoder", f'{"Only encoder is supported yet."}'
-
     output_size = verify_layers_io(input_size, architecture)
 
-    input_layer = create_input_layer(
+    input_layer = build_input_layer(
         input_size, padding_idx=padding_idx, **architecture[0]
     )
 
@@ -370,9 +285,9 @@ def create_network(
         layer_args = {x: layer_conf[x] for x in layer_conf if x != "layer_type"}
 
         if layer_type == "transformer":
-            module = create_transformer_layer(**layer_args)
+            module = build_transformer_layer(**layer_args)
         elif layer_type == "conformer":
-            module = create_conformer_layer(**layer_args)
+            module = build_conformer_layer(**layer_args)
 
         architecture_modules.append(module)
